+++ math/knot/knot_vassiliev.cpp
@@ -33,8 +33,8 @@ void vector_product(const double *v1, const double *v2, double *m) {
 #define			next			Parent -> next
 #define			prev			Parent -> prev
 
-float knot::prmAcn::compute() {
-	float value = 0.0;
+double knot::prmAcn::compute() {
+	double value = 0.0;
 
 	std::size_t i1, i2;
 
@@ -73,8 +73,8 @@ float knot::prmAcn::compute() {
 	return value / (2 * M_PI);
 }
 
-float knot::prmSAcn::compute() {
-	float value = 0.0;
+double knot::prmSAcn::compute() {
+	double value = 0.0;
 
 	std::size_t i1, i2;
 
@@ -113,8 +113,8 @@ float knot::prmSAcn::compute() {
 	return value / (2 * M_PI);
 }
 
-float knot::prmKI::compute() {
-	float value = 0.0;
+double knot::prmKI::compute() {
+	double value = 0.0;
 
 	std::size_t i1, i2;
 	int o;
@@ -195,8 +195,8 @@ float knot::prmKI::compute() {
 	return value / (4 * M_PI * M_PI);
 }
 
-//float knot::prmExperimental::compute() {
-//	float value = 0.0;
+//double knot::prmExperimental::compute() {
+//	double value = 0.0;
 //
 //	std::size_t i1, i2, i3, i4;
 //
@@ -286,8 +286,8 @@ float knot::prmKI::compute() {
 //	return value / (4 * M_PI * M_PI);
 //}
 
-float knot::prmExperimental::compute() {
-	float value = 0.0;
+double knot::prmExperimental::compute() {
+	double value = 0.0;
 
 	std::size_t i1, i2;
 
@@ -414,8 +414,8 @@ static int intersected (const double *a,
 	return 1;
 }
 
-float knot::prmSingular::compute() {
-	float value = 0.0;
+double knot::prmSingular::compute() {
+	double value = 0.0;
 
 	std::size_t i1, i2, i3, i4;
 
@@ -530,8 +530,8 @@ float knot::prmSingular::compute() {
 	return value;
 }
 
-float knot::prmExperimental2::compute() {
-	float value = 0.0;
+double knot::prmExperimental2::compute() {
+	double value = 0.0;
 
 	std::size_t i1, i2;
 	int o;
